test: move personalization tests to separate file
